requirements.txt
Django==1.7
django-debug-toolbar==1.2.1
latimes-calculate==0.2.30
python-dateutil==2.2
python-memcached==1.53
django_tables2
django-bootstrap-toolkit
